[Synthetics] Update monitor details page loading (#145931)
(cherry picked from commit 698504708a48b4706fbe2d4cb98bd2f997da3d6d)
