+++ x-pack/plugins/synthetics/common/runtime_types/monitor_management/monitor_types.ts
@@ -349,7 +349,7 @@ export const EncryptedSyntheticsMonitorWithIdCodec = t.intersection([
 // TODO: Remove EncryptedSyntheticsMonitorWithIdCodec (as well as SyntheticsMonitorWithIdCodec if possible) along with respective TypeScript types in favor of EncryptedSyntheticsSavedMonitorCodec
 export const EncryptedSyntheticsSavedMonitorCodec = t.intersection([
   EncryptedSyntheticsMonitorCodec,
-  t.interface({ id: t.string, updated_at: t.string }),
+  t.interface({ id: t.string, updated_at: t.string, created_at: t.string }),
 ]);
 
 export type SyntheticsMonitorWithId = t.TypeOf<typeof SyntheticsMonitorWithIdCodec>;
@@ -377,6 +377,7 @@ export const MonitorManagementListResultCodec = t.type({
       id: t.string,
       attributes: EncryptedSyntheticsMonitorCodec,
       updated_at: t.string,
+      created_at: t.string,
     })
   ),
   page: t.number,
