+++ x-pack/plugins/synthetics/public/apps/synthetics/state/monitor_details/api.ts
@@ -76,5 +76,6 @@ export const fetchSyntheticsMonitor = async ({
   return {
     ...savedObject.attributes,
     updated_at: savedObject.updated_at,
+    created_at: savedObject.created_at,
   } as EncryptedSyntheticsSavedMonitor;
 };
