+++ x-pack/plugins/synthetics/public/legacy_uptime/components/monitor_management/hooks/use_invalid_monitors.tsx
@@ -40,6 +40,7 @@ export const useInvalidMonitors = (errorSummaries?: Ping[]) => {
           .map(({ saved_object: savedObject }) => ({
             ...savedObject,
             updated_at: savedObject.updatedAt!,
+            created_at: savedObject.createdAt!,
           }));
       }
     }
