+++ x-pack/plugins/synthetics/public/legacy_uptime/components/monitor_management/monitor_list/monitor_list.test.tsx
@@ -28,6 +28,7 @@ describe('<MonitorManagementList />', () => {
     monitors.push({
       id: `test-monitor-id-${i}`,
       updated_at: '123',
+      created_at: '123',
       attributes: {
         config_id: `test-monitor-id-${i}`,
         name: `test-monitor-${i}`,
